+++ document/format.go
@@ -1,10 +1,8 @@
 package document
 
-import "text/template"
-
 type Formatter interface {
 	ToFilePath(path string, kind string) (string, error)
 	ToLinkPath(path string, kind string) (string, error)
 	ProcessMarkdown(name, summary, text string) (string, error)
-	WriteAuxiliary(p *Package, dir string, t *template.Template) error
+	WriteAuxiliary(p *Package, dir string, proc *Processor) error
 }