+++ document/links.go
@@ -29,64 +29,93 @@ func findLinks(text string) ([]int, error) {
 	return links, nil
 }
 
-func (proc *Processor) ProcessLinks(doc *Docs) error {
-	lookup := proc.collectPaths(doc)
-	//for k, v := range lookup {
-	//	fmt.Println(k, v.Elements)
-	//}
-	return proc.processLinksPackage(doc.Decl, []string{}, lookup)
+func (proc *Processor) ProcessLinks() error {
+	proc.filterPackages()
+	proc.collectPaths()
+
+	if !proc.UseExports {
+		for k := range proc.linkTargets {
+			proc.linkExports[k] = k
+		}
+	}
+	if err := proc.processLinksPackage(proc.Docs.Decl, []string{}, true); err != nil {
+		return err
+	}
+	if err := proc.processLinksPackage(proc.ExportDocs.Decl, []string{}, false); err != nil {
+		return err
+	}
+	return nil
 }
 
-func (proc *Processor) processLinksPackage(p *Package, elems []string, lookup map[string]elemPath) error {
+func (proc *Processor) processLinksPackage(p *Package, elems []string, firstPass bool) error {
 	newElems := appendNew(elems, p.GetName())
 
 	var err error
-	p.Summary, err = proc.replaceLinks(p.Summary, newElems, len(newElems), lookup)
+	p.Summary, err = proc.replaceLinks(p.Summary, newElems, len(newElems), firstPass)
 	if err != nil {
 		return err
 	}
-	p.Description, err = proc.replaceLinks(p.Description, newElems, len(newElems), lookup)
+	p.Description, err = proc.replaceLinks(p.Description, newElems, len(newElems), firstPass)
 	if err != nil {
 		return err
 	}
 
 	for _, pkg := range p.Packages {
-		proc.processLinksPackage(pkg, newElems, lookup)
+		proc.processLinksPackage(pkg, newElems, firstPass)
 	}
 	for _, mod := range p.Modules {
-		proc.processLinksModule(mod, newElems, lookup)
+		proc.processLinksModule(mod, newElems, firstPass)
+	}
+
+	for _, f := range p.Functions {
+		err := proc.processLinksFunction(f, newElems, firstPass)
+		if err != nil {
+			return err
+		}
+	}
+	for _, s := range p.Structs {
+		err := proc.processLinksStruct(s, newElems, firstPass)
+		if err != nil {
+			return err
+		}
+	}
+	for _, tr := range p.Traits {
+		err := proc.processLinksTrait(tr, newElems, firstPass)
+		if err != nil {
+			return err
+		}
 	}
 
 	return nil
 }
 
-func (proc *Processor) processLinksModule(m *Module, elems []string, lookup map[string]elemPath) error {
+func (proc *Processor) processLinksModule(m *Module, elems []string, firstPass bool) error {
 	newElems := appendNew(elems, m.GetName())
 
 	var err error
-	m.Summary, err = proc.replaceLinks(m.Summary, newElems, len(newElems), lookup)
+	m.Summary, err = proc.replaceLinks(m.Summary, newElems, len(newElems), firstPass)
 	if err != nil {
 		return err
 	}
-	m.Description, err = proc.replaceLinks(m.Description, newElems, len(newElems), lookup)
+	m.Description, err = proc.replaceLinks(m.Description, newElems, len(newElems), firstPass)
 	if err != nil {
 		return err
 	}
 
 	for _, f := range m.Functions {
-		err := proc.processLinksFunction(f, newElems, lookup)
+		err := proc.processLinksFunction(f, newElems, firstPass)
 		if err != nil {
 			return err
 		}
 	}
 	for _, s := range m.Structs {
-		err := proc.processLinksStruct(s, newElems, lookup)
+		err := proc.processLinksStruct(s, newElems, firstPass)
 		if err != nil {
 			return err
 		}
 	}
 	for _, tr := range m.Traits {
-		err := proc.processLinksTrait(tr, newElems, lookup)
+		err := proc.processLinksTrait(tr, newElems, firstPass)
 		if err != nil {
 			return err
 		}
@@ -95,37 +124,37 @@ func (proc *Processor) processLinksModule(m *Module, elems []string, lookup map[
 	return nil
 }
 
-func (proc *Processor) processLinksStruct(s *Struct, elems []string, lookup map[string]elemPath) error {
+func (proc *Processor) processLinksStruct(s *Struct, elems []string, firstPass bool) error {
 	newElems := appendNew(elems, s.GetName())
 
 	var err error
-	s.Summary, err = proc.replaceLinks(s.Summary, newElems, len(elems), lookup)
+	s.Summary, err = proc.replaceLinks(s.Summary, newElems, len(elems), firstPass)
 	if err != nil {
 		return err
 	}
-	s.Description, err = proc.replaceLinks(s.Description, newElems, len(elems), lookup)
+	s.Description, err = proc.replaceLinks(s.Description, newElems, len(elems), firstPass)
 	if err != nil {
 		return err
 	}
 
 	for _, p := range s.Parameters {
-		p.Description, err = proc.replaceLinks(p.Description, newElems, len(elems), lookup)
+		p.Description, err = proc.replaceLinks(p.Description, newElems, len(elems), firstPass)
 		if err != nil {
 			return err
 		}
 	}
 	for _, f := range s.Fields {
-		f.Summary, err = proc.replaceLinks(f.Summary, newElems, len(elems), lookup)
+		f.Summary, err = proc.replaceLinks(f.Summary, newElems, len(elems), firstPass)
 		if err != nil {
 			return err
 		}
-		f.Description, err = proc.replaceLinks(f.Description, newElems, len(elems), lookup)
+		f.Description, err = proc.replaceLinks(f.Description, newElems, len(elems), firstPass)
 		if err != nil {
 			return err
 		}
 	}
 	for _, f := range s.Functions {
-		if err := proc.processLinksMethod(f, elems, lookup); err != nil {
+		if err := proc.processLinksMethod(f, elems, firstPass); err != nil {
 			return err
 		}
 	}
@@ -133,15 +162,15 @@ func (proc *Processor) processLinksStruct(s *Struct, elems []string, lookup map[
 	return nil
 }
 
-func (proc *Processor) processLinksTrait(tr *Trait, elems []string, lookup map[string]elemPath) error {
+func (proc *Processor) processLinksTrait(tr *Trait, elems []string, firstPass bool) error {
 	newElems := appendNew(elems, tr.GetName())
 
 	var err error
-	tr.Summary, err = proc.replaceLinks(tr.Summary, newElems, len(elems), lookup)
+	tr.Summary, err = proc.replaceLinks(tr.Summary, newElems, len(elems), firstPass)
 	if err != nil {
 		return err
 	}
-	tr.Description, err = proc.replaceLinks(tr.Description, newElems, len(elems), lookup)
+	tr.Description, err = proc.replaceLinks(tr.Description, newElems, len(elems), firstPass)
 	if err != nil {
 		return err
 	}
@@ -154,17 +183,17 @@ func (proc *Processor) processLinksTrait(tr *Trait, elems []string, lookup map[s
 		}
 	}*/
 	for _, f := range tr.Fields {
-		f.Summary, err = proc.replaceLinks(f.Summary, newElems, len(elems), lookup)
+		f.Summary, err = proc.replaceLinks(f.Summary, newElems, len(elems), firstPass)
 		if err != nil {
 			return err
 		}
-		f.Description, err = proc.replaceLinks(f.Description, newElems, len(elems), lookup)
+		f.Description, err = proc.replaceLinks(f.Description, newElems, len(elems), firstPass)
 		if err != nil {
 			return err
 		}
 	}
 	for _, f := range tr.Functions {
-		if err := proc.processLinksMethod(f, elems, lookup); err != nil {
+		if err := proc.processLinksMethod(f, elems, firstPass); err != nil {
 			return err
 		}
 	}
@@ -172,42 +201,42 @@ func (proc *Processor) processLinksTrait(tr *Trait, elems []string, lookup map[s
 	return nil
 }
 
-func (proc *Processor) processLinksFunction(f *Function, elems []string, lookup map[string]elemPath) error {
+func (proc *Processor) processLinksFunction(f *Function, elems []string, firstPass bool) error {
 	newElems := appendNew(elems, f.GetName())
 
 	var err error
-	f.Summary, err = proc.replaceLinks(f.Summary, newElems, len(elems), lookup)
+	f.Summary, err = proc.replaceLinks(f.Summary, newElems, len(elems), firstPass)
 	if err != nil {
 		return err
 	}
-	f.Description, err = proc.replaceLinks(f.Description, newElems, len(elems), lookup)
+	f.Description, err = proc.replaceLinks(f.Description, newElems, len(elems), firstPass)
 	if err != nil {
 		return err
 	}
-	f.ReturnsDoc, err = proc.replaceLinks(f.ReturnsDoc, newElems, len(elems), lookup)
+	f.ReturnsDoc, err = proc.replaceLinks(f.ReturnsDoc, newElems, len(elems), firstPass)
 	if err != nil {
 		return err
 	}
-	f.RaisesDoc, err = proc.replaceLinks(f.RaisesDoc, newElems, len(elems), lookup)
+	f.RaisesDoc, err = proc.replaceLinks(f.RaisesDoc, newElems, len(elems), firstPass)
 	if err != nil {
 		return err
 	}
 
 	for _, a := range f.Args {
-		a.Description, err = proc.replaceLinks(a.Description, newElems, len(elems), lookup)
+		a.Description, err = proc.replaceLinks(a.Description, newElems, len(elems), firstPass)
 		if err != nil {
 			return err
 		}
 	}
 	for _, p := range f.Parameters {
-		p.Description, err = proc.replaceLinks(p.Description, newElems, len(elems), lookup)
+		p.Description, err = proc.replaceLinks(p.Description, newElems, len(elems), firstPass)
 		if err != nil {
 			return err
 		}
 	}
 
 	for _, o := range f.Overloads {
-		err := proc.processLinksFunction(o, elems, lookup)
+		err := proc.processLinksFunction(o, elems, firstPass)
 		if err != nil {
 			return err
 		}
@@ -216,40 +245,40 @@ func (proc *Processor) processLinksFunction(f *Function, elems []string, lookup 
 	return nil
 }
 
-func (proc *Processor) processLinksMethod(f *Function, elems []string, lookup map[string]elemPath) error {
+func (proc *Processor) processLinksMethod(f *Function, elems []string, firstPass bool) error {
 	var err error
-	f.Summary, err = proc.replaceLinks(f.Summary, elems, len(elems), lookup)
+	f.Summary, err = proc.replaceLinks(f.Summary, elems, len(elems), firstPass)
 	if err != nil {
 		return err
 	}
-	f.Description, err = proc.replaceLinks(f.Description, elems, len(elems), lookup)
+	f.Description, err = proc.replaceLinks(f.Description, elems, len(elems), firstPass)
 	if err != nil {
 		return err
 	}
-	f.ReturnsDoc, err = proc.replaceLinks(f.ReturnsDoc, elems, len(elems), lookup)
+	f.ReturnsDoc, err = proc.replaceLinks(f.ReturnsDoc, elems, len(elems), firstPass)
 	if err != nil {
 		return err
 	}
-	f.RaisesDoc, err = proc.replaceLinks(f.RaisesDoc, elems, len(elems), lookup)
+	f.RaisesDoc, err = proc.replaceLinks(f.RaisesDoc, elems, len(elems), firstPass)
 	if err != nil {
 		return err
 	}
 
 	for _, a := range f.Args {
-		a.Description, err = proc.replaceLinks(a.Description, elems, len(elems), lookup)
+		a.Description, err = proc.replaceLinks(a.Description, elems, len(elems), firstPass)
 		if err != nil {
 			return err
 		}
 	}
 	for _, p := range f.Parameters {
-		p.Description, err = proc.replaceLinks(p.Description, elems, len(elems), lookup)
+		p.Description, err = proc.replaceLinks(p.Description, elems, len(elems), firstPass)
 		if err != nil {
 			return err
 		}
 	}
 
 	for _, o := range f.Overloads {
-		err := proc.processLinksMethod(o, elems, lookup)
+		err := proc.processLinksMethod(o, elems, firstPass)
 		if err != nil {
 			return err
 		}
@@ -258,7 +287,14 @@ func (proc *Processor) processLinksMethod(f *Function, elems []string, lookup ma
 	return nil
 }
 
-func (proc *Processor) replaceLinks(text string, elems []string, modElems int, lookup map[string]elemPath) (string, error) {
+func (proc *Processor) replaceLinks(text string, elems []string, modElems int, firstPass bool) (string, error) {
+	if firstPass {
+		return proc.replaceLinksFirstPass(text, elems, modElems)
+	}
+	return proc.replaceLinksSecondPass(text, elems, modElems)
+}
+
+func (proc *Processor) replaceLinksSecondPass(text string, elems []string, modElems int) (string, error) {
 	indices, err := findLinks(text)
 	if err != nil {
 		return "", err
@@ -270,7 +306,7 @@ func (proc *Processor) replaceLinks(text string, elems []string, modElems int, l
 		start, end := indices[i], indices[i+1]
 		link := text[start+1 : end-1]
 
-		entry, linkText, parts, ok := toLink(link, elems, modElems, lookup, proc.ShortLinks)
+		entry, linkText, parts, ok := placeholderToLink(link, elems, modElems, proc.linkTargets, proc.ShortLinks)
 		if !ok {
 			continue
 		}
@@ -281,6 +317,7 @@ func (proc *Processor) replaceLinks(text string, elems []string, modElems int, l
 		} else {
 			basePath = path.Join(parts...)
 		}
+
 		pathStr, err := proc.Formatter.ToLinkPath(basePath, entry.Kind)
 		if err != nil {
 			return "", err
@@ -293,13 +330,30 @@ func (proc *Processor) replaceLinks(text string, elems []string, modElems int, l
 	return text, nil
 }
 
-func toLink(link string, elems []string, modElems int, lookup map[string]elemPath, shorten bool) (entry *elemPath, text string, parts []string, ok bool) {
-	linkParts := strings.SplitN(link, " ", 2)
-	if strings.HasPrefix(link, ".") {
-		entry, text, parts, ok = toRelLink(linkParts[0], elems, modElems, lookup)
-	} else {
-		entry, text, parts, ok = toAbsLink(linkParts[0], elems, modElems, lookup)
+func (proc *Processor) replaceLinksFirstPass(text string, elems []string, modElems int) (string, error) {
+	indices, err := findLinks(text)
+	if err != nil {
+		return "", err
 	}
+	if len(indices) == 0 {
+		return text, nil
+	}
+	for i := len(indices) - 2; i >= 0; i -= 2 {
+		start, end := indices[i], indices[i+1]
+		link := text[start+1 : end-1]
+
+		content, ok := linkToPlaceholder(link, elems, modElems, proc.linkExports)
+		if !ok {
+			continue
+		}
+		text = fmt.Sprintf("%s[%s]%s", text[:start], content, text[end:])
+	}
+	return text, nil
+}
+
+func placeholderToLink(link string, elems []string, modElems int, lookup map[string]elemPath, shorten bool) (entry *elemPath, text string, parts []string, ok bool) {
+	linkParts := strings.SplitN(link, " ", 2)
+	entry, text, parts, ok = placeholderToAbsLink(linkParts[0], elems, modElems, lookup)
 	if !ok {
 		return
 	}
@@ -319,39 +373,7 @@ func toLink(link string, elems []string, modElems int, lookup map[string]elemPat
 	return
 }
 
-func toRelLink(link string, elems []string, modElems int, lookup map[string]elemPath) (*elemPath, string, []string, bool) {
-	dots := 0
-	fullPath := []string{}
-	for strings.HasPrefix(link[dots:], ".") {
-		if dots > 0 {
-			fullPath = append(fullPath, "..")
-		}
-		dots++
-	}
-	if dots > modElems {
-		log.Printf("WARNING: Too many leading dots in cross ref '%s' in %s", link, strings.Join(elems, "."))
-		return nil, "", nil, false
-	}
-	linkText := link[dots:]
-	subElems := elems[:modElems-(dots-1)]
-	var fullLink string
-	if len(subElems) == 0 {
-		fullLink = linkText
-	} else {
-		fullLink = strings.Join(subElems, ".") + "." + linkText
-	}
-
-	elemPath, ok := lookup[fullLink]
-	if !ok {
-		log.Printf("WARNING: Can't resolve cross ref '%s' (%s) in %s", link, fullLink, strings.Join(elems, "."))
-		return nil, "", nil, false
-	}
-
-	fullPath = append(fullPath, elemPath.Elements[len(subElems):]...)
-	return &elemPath, linkText, fullPath, true
-}
-
-func toAbsLink(link string, elems []string, modElems int, lookup map[string]elemPath) (*elemPath, string, []string, bool) {
+func placeholderToAbsLink(link string, elems []string, modElems int, lookup map[string]elemPath) (*elemPath, string, []string, bool) {
 	elemPath, ok := lookup[link]
 	if !ok {
 		log.Printf("WARNING: Can't resolve cross ref '%s' in %s", link, strings.Join(elems, "."))
@@ -374,4 +396,59 @@ func toAbsLink(link string, elems []string, modElems int, lookup map[string]elem
 	}
 	fullPath = append(fullPath, elemPath.Elements[skip:]...)
 	return &elemPath, link, fullPath, true
+}
+
+func linkToPlaceholder(link string, elems []string, modElems int, lookup map[string]string) (string, bool) {
+	linkParts := strings.SplitN(link, " ", 2)
+
+	var placeholder string
+	var ok bool
+	if strings.HasPrefix(link, ".") {
+		placeholder, ok = linkToPlaceholderRel(linkParts[0], elems, modElems, lookup)
+	} else {
+		placeholder, ok = linkToPlaceholderAbs(linkParts[0], elems, lookup)
+	}
+	if !ok {
+		return "", false
+	}
+	if len(linkParts) > 1 {
+		return fmt.Sprintf("%s %s", placeholder, linkParts[1]), true
+	} else {
+		return placeholder, true
+	}
+}
+
+func linkToPlaceholderRel(link string, elems []string, modElems int, lookup map[string]string) (string, bool) {
+	dots := 0
+	for strings.HasPrefix(link[dots:], ".") {
+		dots++
+	}
+	if dots > modElems {
+		log.Printf("WARNING: Too many leading dots in cross ref '%s' in %s", link, strings.Join(elems, "."))
+		return "", false
+	}
+	linkText := link[dots:]
+	subElems := elems[:modElems-(dots-1)]
+	var fullLink string
+	if len(subElems) == 0 {
+		fullLink = linkText
+	} else {
+		fullLink = strings.Join(subElems, ".") + "." + linkText
+	}
+
+	placeholder, ok := lookup[fullLink]
+	if !ok {
+		log.Printf("WARNING: Can't resolve cross ref '%s' (%s) in %s", link, fullLink, strings.Join(elems, "."))
+		return "", false
+	}
+	return placeholder, true
+}
+
+func linkToPlaceholderAbs(link string, elems []string, lookup map[string]string) (string, bool) {
+	placeholder, ok := lookup[link]
+	if !ok {
+		log.Printf("WARNING: Can't resolve cross ref '%s' in %s", link, strings.Join(elems, "."))
+		return "", false
+	}
+	return placeholder, true
 }