+++ document/paths_test.go
@@ -82,9 +82,10 @@ func TestCollectPaths(t *testing.T) {
 		},
 	}
 
-	proc := NewProcessor(nil, nil, false)
-	p := proc.collectPaths(&docs)
-	assert.Equal(t, 11, len(p))
+	proc := NewProcessor(&docs, nil, nil, false, false)
+	proc.ExportDocs = &docs
+	proc.collectPaths()
+	assert.Equal(t, 11, len(proc.linkTargets))
 
 	tests := []struct {
 		mem string
@@ -99,7 +100,7 @@ func TestCollectPaths(t *testing.T) {
 	}
 
 	for _, tt := range tests {
-		obs, ok := p[tt.mem]
+		obs, ok := proc.linkTargets[tt.mem]
 		assert.Equal(t, tt.ok, ok)
 		assert.Equal(t, tt.exp, obs.Elements)
 	}